+++ RELEASENOTES.md
@@ -1,5 +1,20 @@
 # Release notes of ion-autocomplete
 
+## Version 0.3.2
+
+* Tag: [0.3.2](https://github.com/guylabs/ion-autocomplete/tree/v0.3.2)
+* Release: [ion-autocomplete-0.3.2.zip](https://github.com/guylabs/ion-autocomplete/archive/v0.3.2.zip)
+
+### Changes
+
+* A single selected item is not an array anymore with a single value - [#115](https://github.com/guylabs/ion-autocomplete/issues/115)
+* The item repeat has been switched from `collection-repeat` to `ng-repeat` as there were some issues - [#126](https://github.com/guylabs/ion-autocomplete/issues/126)
+
+### Migration notes
+
+* As part of the [#115](https://github.com/guylabs/ion-autocomplete/issues/115) issue when you now select a single value (`maxSelectedItems=1`) then the item
+is now returned as object and not as an array with one element. If you use a custom template, please also check the changes in the default template.
+
 ## Version 0.3.1
 
 * Tag: [0.3.1](https://github.com/guylabs/ion-autocomplete/tree/v0.3.1)
@@ -7,15 +22,18 @@
 
 ### Changes
 
-* Added possibility to pre populate the search items - [#57](https://github.com/guylabs/ion-autocomplete/issues/57).
 * Added the `max-selected-items` attribute to restrict the selected search items - [#58](https://github.com/guylabs/ion-autocomplete/issues/58).
 * Added the `cancel-button-clicked-method` attribute to be able to get notified when the cancel button is clicked - [#63](https://github.com/guylabs/ion-autocomplete/issues/63).
 * Added the `external-model` attribute to be able to prepopulate the selected items and to clear them programmatically - [#66](https://github.com/guylabs/ion-autocomplete/issues/66), [#89](https://github.com/guylabs/ion-autocomplete/issues/89).
+* Added the ability to pass the `ng-model-options` to the inner search input field - [#91](https://github.com/guylabs/ion-autocomplete/issues/91)
+* Added the ability to initialize the search items within the `items-method` - [#57](https://github.com/guylabs/ion-autocomplete/issues/57)
 
 ### Migration notes
 
 * As of version `0.3.1` the `multiple-select` attribute has been dropped in favor of the `max-selected-items` attribute.
 Please have a look at the documentation here https://github.com/guylabs/ion-autocomplete#the-max-selected-items on how to migrate this.
+* The `search-items` attribute has been removed as now the initialization of the `search-items` is done in the `items-method`. See the 
+new documentation here https://github.com/guylabs/ion-autocomplete#the-items-method.
 
 ## Version 0.3.0